add code

+++ kompilator.sh
@@ -8,30 +8,22 @@ read KERNEL
 echo "Podaj starą wersję kernela np. 5.5"
 read SKERNEL
 
-KERNEL_EXIST="linux-${SKERNEL}/.config"
+SKERNEL_EXIST="linux-${SKERNEL}/.config"
+KERNEL_EXIST="linux-${KERNEL}.tar.xz"
 
-if [ -e $KERNEL_EXIST ]
+if [ -e $SKERNEL_EXIST ] && [ $KERNEL_EXIST ]
 	then
 	
-	unxz -c linux-${KERNEL}.tar.xz | gpg --verify linux-${KERNEL}.tar.sign -
-	pause
+	unxz -c linux-${KERNEL}.tar.xz | gpg --verify linux-${KERNEL}.tar.sign -	
 	tar xavf linux-${KERNEL}.tar.xz
-	
-	elif [ -e linux-${SKERNEL}/.config]
-	
+	cp linux-${SKERNEL}/.config linux-${KERNEL}/.config
+	cd linux-${KERNEL}
+	make menuconfig
+	make -j 16
+	sudo make modules_install
+	sudo cp -v arch/x86_64/boot/bzImage /boot/vmlinuz-linux-${KERNEL}
 
-		cp linux-${SKERNEL}/.config linux-${KERNEL}/.config
-
-		cd linux-${KERNEL}
-
-		make menuconfig
-
-		make -j 16
-
-		sudo make modules_install
-
-		sudo cp -v arch/x86_64/boot/bzImage /boot/vmlinuz-linux-${KERNEL}
 	else
-		echo "Brak plików"
+	echo "Brak plików"
 fi
